internal/gocdk: fold the "launch" command into "deploy"
