+++ internal/cmd/gocdk/testdata/recordreplay/BuildHelp/commands.txt
@@ -1,0 +1,3 @@
+gocdk help build
+gocdk build -h
+gocdk build --help
