+++ internal/cmd/gocdk/testdata/recordreplay/BuildHelp/out.txt
@@ -1,0 +1,33 @@
+$ gocdk help build
+TODO more about build
+
+Usage:
+  gocdk build [flags]
+
+Flags:
+  -h, --help                     help for build
+      --list                     display Docker images of this project
+  -t, --tag name[:tag] OR :tag   name and/or tag in the form name[:tag] OR :tag
+
+$ gocdk build -h
+TODO more about build
+
+Usage:
+  gocdk build [flags]
+
+Flags:
+  -h, --help                     help for build
+      --list                     display Docker images of this project
+  -t, --tag name[:tag] OR :tag   name and/or tag in the form name[:tag] OR :tag
+
+$ gocdk build --help
+TODO more about build
+
+Usage:
+  gocdk build [flags]
+
+Flags:
+  -h, --help                     help for build
+      --list                     display Docker images of this project
+  -t, --tag name[:tag] OR :tag   name and/or tag in the form name[:tag] OR :tag
+
