+++ internal/cmd/gocdk/testdata/recordreplay/Demo/commands.txt
@@ -1,3 +1,5 @@
+gocdk demo list
+
 # Add fails when we're not in a module.
 gocdk demo add blob --> FAIL
