+++ internal/cmd/gocdk/testdata/recordreplay/Demo/out.txt
@@ -1,3 +1,10 @@
+$ gocdk demo list
+blob
+docstore
+pubsub
+runtimevar
+secrets
+
 $ gocdk demo add blob
 Error: demo add: couldn't find a Go module root at or above [ROOTDIR]
