--- internal/cmd/gocdk/testdata/recordreplay/DemoList/commands.txt
@@ -1,1 +1,0 @@
-gocdk demo list
