--- internal/cmd/gocdk/testdata/recordreplay/DemoList/out.txt
@@ -1,7 +1,0 @@
-$ gocdk demo list
-blob
-docstore
-pubsub
-runtimevar
-secrets
-
