+++ internal/cmd/gocdk/testdata/recordreplay/Deploy/commands.txt
@@ -1,0 +1,2 @@
+# Deploy fails when we're not in a module.
+gocdk deploy --> FAIL
