+++ internal/cmd/gocdk/testdata/recordreplay/DeployHelp/commands.txt
@@ -1,0 +1,3 @@
+gocdk help deploy
+gocdk deploy -h
+gocdk deploy --help
