+++ internal/cmd/gocdk/testdata/recordreplay/DeployHelp/out.txt
@@ -1,0 +1,33 @@
+$ gocdk help deploy
+TODO more about deploy
+
+Usage:
+  gocdk deploy BIOME [flags]
+
+Flags:
+      --apply biome apply   whether to run biome apply before deploying (default true)
+  -h, --help                help for deploy
+      --image name[:tag]    Docker image to deploy in the form name[:tag] OR `:tag`, or empty string to generate a tag (default ":latest")
+
+$ gocdk deploy -h
+TODO more about deploy
+
+Usage:
+  gocdk deploy BIOME [flags]
+
+Flags:
+      --apply biome apply   whether to run biome apply before deploying (default true)
+  -h, --help                help for deploy
+      --image name[:tag]    Docker image to deploy in the form name[:tag] OR `:tag`, or empty string to generate a tag (default ":latest")
+
+$ gocdk deploy --help
+TODO more about deploy
+
+Usage:
+  gocdk deploy BIOME [flags]
+
+Flags:
+      --apply biome apply   whether to run biome apply before deploying (default true)
+  -h, --help                help for deploy
+      --image name[:tag]    Docker image to deploy in the form name[:tag] OR `:tag`, or empty string to generate a tag (default ":latest")
+
