+++ internal/cmd/gocdk/testdata/recordreplay/Help/out.txt
@@ -11,7 +11,6 @@ Available Commands:
   deploy      TODO Deploy the biome
   help        Help about any command
   init        TODO: Initialize a new project
-  launch      TODO Launch BIOME
   resource    TODO Resources
   serve       TODO Run an auto-reloading local server
 
@@ -33,7 +32,6 @@ Available Commands:
   deploy      TODO Deploy the biome
   help        Help about any command
   init        TODO: Initialize a new project
-  launch      TODO Launch BIOME
   resource    TODO Resources
   serve       TODO Run an auto-reloading local server
 
@@ -55,7 +53,6 @@ Available Commands:
   deploy      TODO Deploy the biome
   help        Help about any command
   init        TODO: Initialize a new project
-  launch      TODO Launch BIOME
   resource    TODO Resources
   serve       TODO Run an auto-reloading local server
 
@@ -77,7 +74,6 @@ Available Commands:
   deploy      TODO Deploy the biome
   help        Help about any command
   init        TODO: Initialize a new project
-  launch      TODO Launch BIOME
   resource    TODO Resources
   serve       TODO Run an auto-reloading local server
