+++ internal/cmd/gocdk/testdata/recordreplay/Resource/commands.txt
@@ -1,3 +1,5 @@
+gocdk resource list
+
 # Add fails when we're not in a module.
 gocdk resource add dev blob/fileblob --> FAIL
