--- internal/cmd/gocdk/testdata/recordreplay/ResourceList/commands.txt
@@ -1,1 +1,0 @@
-gocdk resource list
