--- internal/cmd/gocdk/testdata/recordreplay/ResourceList/out.txt
@@ -1,18 +1,0 @@
-$ gocdk resource list
-blob/azureblob
-blob/fileblob
-blob/gcsblob
-blob/s3blob
-docstore/awsdynamodb
-docstore/azurecosmos
-docstore/gcpfirestore
-pubsub/awssnssqs
-pubsub/azuresb
-pubsub/gcppubsub
-runtimevar/awsparamstore
-runtimevar/filevar
-runtimevar/gcpruntimeconfig
-secrets/awskms
-secrets/azurekeyvault
-secrets/gcpkms
-
