+++ internal/cmd/gocdk/testdata/recordreplay/Serve/out.txt
@@ -1,0 +1,19 @@
+$ gocdk serve
+Error: gocdk serve: couldn't find a Go module root at or above [ROOTDIR]
+
+$ gocdk init myproj
+gocdk: Project created at [ROOTDIR]/myproj with:
+gocdk: - Go HTTP server
+gocdk: - Dockerfile
+gocdk: - 'dev' biome for local development settings
+gocdk: Run `cd myproj`, then run:
+gocdk: - `gocdk serve` to run the server locally with live code reloading
+gocdk: - `gocdk demo` to test new APIs
+gocdk: - `gocdk build` to build a Docker container
+gocdk: - `gocdk biome add` to configure launch settings
+
+$ cd myproj
+
+$ gocdk serve --biome=foo
+Error: gocdk serve: biome foo not found
+
