+++ internal/cmd/gocdk/testdata/recordreplay/ServeHelp/commands.txt
@@ -1,0 +1,3 @@
+gocdk help serve
+gocdk serve -h
+gocdk serve --help
