+++ internal/cmd/gocdk/testdata/recordreplay/ServeHelp/out.txt
@@ -1,0 +1,36 @@
+$ gocdk help serve
+TODO more about serve
+
+Usage:
+  gocdk serve [flags]
+
+Flags:
+      --address host:port        host:port address to serve on (default "localhost:8080")
+      --biome name               name of biome to apply and use configuration from (default "dev")
+  -h, --help                     help for serve
+      --poll-interval duration   time between checking project directory for changes (default 500ms)
+
+$ gocdk serve -h
+TODO more about serve
+
+Usage:
+  gocdk serve [flags]
+
+Flags:
+      --address host:port        host:port address to serve on (default "localhost:8080")
+      --biome name               name of biome to apply and use configuration from (default "dev")
+  -h, --help                     help for serve
+      --poll-interval duration   time between checking project directory for changes (default 500ms)
+
+$ gocdk serve --help
+TODO more about serve
+
+Usage:
+  gocdk serve [flags]
+
+Flags:
+      --address host:port        host:port address to serve on (default "localhost:8080")
+      --biome name               name of biome to apply and use configuration from (default "dev")
+  -h, --help                     help for serve
+      --poll-interval duration   time between checking project directory for changes (default 500ms)
+
